Refactor course page with modular hooks and improved component structure
