+++ app/dashboard/course/[slug]/components/MainContent.tsx
@@ -1,52 +1,35 @@
 "use client"
 
-import type React from "react"
-import { useState, useEffect, useCallback, useMemo, useRef } from "react"
-import { useRouter } from "next/navigation"
-import { useProgress } from "@/hooks"
-import { useToast } from "@/components/ui/use-toast"
-import { Button } from "@/components/ui/button"
-import { 
-  Play, 
-  Lock, 
-  User as UserIcon, 
-  Award, 
-  Badge, 
-  ChevronLeft, 
-  ChevronRight, 
-  Clock, 
-  Maximize2, 
-  Minimize2, 
-  Download, 
-  Share2,
-  PictureInPicture,
-  Settings,
-  RotateCcw
-} from "lucide-react"
-import { useAppDispatch, useAppSelector } from "@/store/hooks"
-import { setCurrentVideoApi, markChapterAsCompleted } from "@/store/slices/course-slice"
-import type { FullCourseType, FullChapterType } from "@/app/types/types"
-import CourseDetailsTabs, { AccessLevels } from "./CourseDetailsTabs"
-import { formatDuration } from "../utils/formatUtils"
-import VideoPlayer from "./video/components/VideoPlayer"
-import VideoNavigationSidebar from "./video/components/VideoNavigationSidebar"
-import { migratedStorage } from "@/lib/storage"
-import AnimatedCourseAILogo from "./video/components/AnimatedCourseAILogo"
-import AutoplayOverlay from "./AutoplayOverlay"
-import VideoGenerationSection from "./VideoGenerationSection"
-import { MarkdownRenderer } from "./markdownUtils"
-import { useVideoState, getVideoBookmarks } from "./video/hooks/useVideoState"
-import { VideoDebug } from "./video/components/VideoDebug"
-import { Card, CardContent } from "@/components/ui/card"
-import { AnimatePresence, motion } from "framer-motion"
-import { useAuth } from "@/modules/auth"
-import { isAdmin } from "@/lib/auth"
-import CourseActions from "./CourseActions"
-import { cn } from "@/lib/utils"
-import { PDFDownloadLink } from "@react-pdf/renderer"
-import CertificateGenerator from "./CertificateGenerator"
-import RecommendedSection from "@/components/shared/RecommendedSection"
-import type { BookmarkData } from "./video/types"
+import React, { useState, useEffect, useCallback, useMemo } from 'react'
+import { useProgress } from '@/hooks'
+import { useAppDispatch, useAppSelector } from '@/store/hooks'
+import { setCurrentVideoApi } from '@/store/slices/course-slice'
+import { useAuth } from '@/modules/auth'
+import { useVideoState } from './video/hooks/useVideoState'
+import { cn } from '@/lib/utils'
+import { formatDuration } from '../utils/formatUtils'
+import { AnimatePresence, motion } from 'framer-motion'
+
+// Custom hooks
+import { useVideoManagement } from '../hooks/useVideoManagement'
+import { useAutoplay } from '../hooks/useAutoplay'
+import { useLayoutManagement } from '../hooks/useLayoutManagement'
+import { useKeyboardShortcuts } from '../hooks/useKeyboardShortcuts'
+
+// Components
+import CourseToolbar from './CourseToolbar'
+import VideoPlayer from './video/components/VideoPlayer'
+import VideoNavigationSidebar from './video/components/VideoNavigationSidebar'
+import CourseDetailsTabs from './CourseDetailsTabs'
+import VideoGenerationSection from './VideoGenerationSection'
+import AutoplayOverlay from './AutoplayOverlay'
+import CertificateGenerator from './CertificateGenerator'
+import RecommendedSection from '@/components/shared/RecommendedSection'
+import VideoDebug from './video/components/VideoDebug'
+
+// Types
+import type { FullCourseType, FullChapterType } from '@/app/types/types'
+import type { BookmarkData } from './video/types'
 
 interface ModernCoursePageProps {
   course: FullCourseType
@@ -57,7 +40,6 @@ interface ModernCoursePageProps {
 const MemoizedVideoNavigationSidebar = React.memo(VideoNavigationSidebar)
 const MemoizedVideoPlayer = React.memo(VideoPlayer)
 const MemoizedCourseDetailsTabs = React.memo(CourseDetailsTabs)
-const MemoizedAnimatedCourseAILogo = React.memo(AnimatedCourseAILogo)
 
 const MainContent: React.FC<ModernCoursePageProps> = ({ 
   course, 
@@ -67,8 +49,6 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
   // ALL HOOKS MUST BE AT THE TOP LEVEL - NO EARLY RETURNS OR CONDITIONS
   // ============================================================================
   
-  const router = useRouter()
-  const { toast } = useToast()
   const dispatch = useAppDispatch()
   const { user, subscription } = useAuth()
   
@@ -80,94 +60,66 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
 
   // Redux state
   const currentVideoId = useAppSelector((state) => state.course.currentVideoId)
-  const courseProgress = useAppSelector((state) => state.course.progress)
 
   // ============================================================================
   // LOCAL STATE DECLARATIONS
   // ============================================================================
   
-  const [sidebarOpen, setSidebarOpen] = useState(false)
-  const [videoEnding, setVideoEnding] = useState(false)
-  const [showCertificate, setShowCertificate] = useState(false)
-  const [resumePromptShown, setResumePromptShown] = useState(false)
   const [videoDurations, setVideoDurations] = useState<Record<string, number>>({})
   const [isVideoLoading, setIsVideoLoading] = useState(true)
-  const [hasPlayedFreeVideo, setHasPlayedFreeVideo] = useState(false)
-  const [showAuthPrompt, setShowAuthPrompt] = useState(false)
-  const [showAutoplayOverlay, setShowAutoplayOverlay] = useState(false)
-  const [showLogoOverlay, setShowLogoOverlay] = useState(false)
-  const [playerRef, setPlayerRef] = useState<React.RefObject<any> | null>(null)
-  const [wideMode, setWideMode] = useState(false)
-  const [isPiPActive, setIsPiPActive] = useState(false)
   const [mobilePlaylistOpen, setMobilePlaylistOpen] = useState(false)
-  const [autoplayMode, setAutoplayMode] = useState(false)
-  const [autoplayCountdown, setAutoplayCountdown] = useState(5)
-  const [nextChapterInfo, setNextChapterInfo] = useState<{
-    title: string
-    description?: string
-    thumbnail?: string
-    duration?: number
-  } | null>(null)
-  const [showChapterTransition, setShowChapterTransition] = useState(false)
-  const [isFullscreen, setIsFullscreen] = useState(false)
-  const [isTheaterMode, setIsTheaterMode] = useState(false)
+  const [showCertificate, setShowCertificate] = useState(false)
+  const [playerRef, setPlayerRef] = useState<React.RefObject<any> | null>(null)
 
   // ============================================================================
-  // REFS AND STORES
+  // CUSTOM HOOKS
   // ============================================================================
   
-  const pipStateRef = useRef(false)
-  const videoStateStore = useVideoState
-  const isOwner = user?.id === course.userId
+  // Video management
+  const { 
+    videoPlaylist, 
+    handleChapterSelect, 
+    handleChapterComplete, 
+    getCurrentChapterInfo 
+  } = useVideoManagement(course, videoDurations)
+
+  // Autoplay functionality
+  const {
+    autoplayMode,
+    autoplayCountdown,
+    nextChapterInfo,
+    showChapterTransition,
+    showAutoplayOverlay,
+    handleAutoplayToggle,
+    handleCancelAutoplay,
+    handleVideoEnd
+  } = useAutoplay()
+
+  // Layout management
+  const {
+    wideMode,
+    isPiPActive,
+    isFullscreen,
+    isTheaterMode,
+    handlePIPToggle,
+    handleWideModeToggle,
+    handleFullscreenToggle,
+    handleTheaterModeToggle,
+    gridLayoutClasses
+  } = useLayoutManagement(course.id)
 
   // ============================================================================
   // COMPUTED VALUES AND MEMOIZED DATA
   // ============================================================================
   
-  // Video playlist with proper typing
-  const videoPlaylist = useMemo(() => {
-    if (!course?.chapters) return []
-    
-    return course.chapters
-      .filter((chapter): chapter is FullChapterType => 
-        Boolean(chapter?.videoId && chapter?.id)
-      )
-      .map((chapter) => ({
-        chapter,
-        videoId: chapter.videoId!,
-        title: chapter.title,
-        description: chapter.description,
-        isFree: chapter.isFree || false,
-        duration: videoDurations[chapter.videoId!] || 0,
-      }))
-      .sort((a, b) => (a.chapter.order || 0) - (b.chapter.order || 0))
-  }, [course?.chapters, videoDurations])
-
-  // Current chapter and index
-  const currentChapter = useMemo(() => {
-    if (!currentVideoId || !videoPlaylist.length) return null
-    return videoPlaylist.find(item => item.videoId === currentVideoId)?.chapter || null
-  }, [currentVideoId, videoPlaylist])
-
-  const currentIndex = useMemo(() => {
-    if (!currentVideoId || !videoPlaylist.length) return -1
-    return videoPlaylist.findIndex(item => item.videoId === currentVideoId)
-  }, [currentVideoId, videoPlaylist])
-
-  // Navigation helpers
-  const nextChapter = useMemo(() => {
-    return currentIndex >= 0 && currentIndex < videoPlaylist.length - 1 
-      ? videoPlaylist[currentIndex + 1] 
-      : null
-  }, [currentIndex, videoPlaylist])
-
-  const prevChapter = useMemo(() => {
-    return currentIndex > 0 ? videoPlaylist[currentIndex - 1] : null
-  }, [currentIndex, videoPlaylist])
-
-  const isLastVideo = useMemo(() => {
-    return currentIndex === videoPlaylist.length - 1
-  }, [currentIndex, videoPlaylist])
+  // Current chapter info
+  const { 
+    currentChapter, 
+    currentIndex, 
+    nextChapter, 
+    prevChapter, 
+    isLastVideo 
+  } = getCurrentChapterInfo(currentVideoId)
 
   // User subscription info
   const userSubscription = useMemo(() => {
@@ -181,63 +133,13 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
     return allowedByChapter || !!userSubscription
   }, [currentChapter?.isFree, userSubscription])
 
-  // Grid layout classes based on PIP state
-  const gridLayoutClasses = useMemo(() => {
-    if (isPiPActive) {
-      return "grid-cols-1" // Single column when PIP is active
-    }
-    return wideMode 
-      ? "md:grid-cols-[1fr_360px] xl:grid-cols-[1fr_420px]" // Wide mode with sidebar
-      : "md:grid-cols-[1fr_360px] xl:grid-cols-[1fr_420px]" // Standard layout
-  }, [isPiPActive, wideMode])
+  // User ownership
+  const isOwner = user?.id === course.userId
 
   // ============================================================================
-  // EVENT HANDLERS - ALL USE CALLBACK TO PREVENT RE-CREATION
+  // EVENT HANDLERS
   // ============================================================================
   
-  const handlePIPToggle = useCallback((isPiPActive: boolean) => {
-    setIsPiPActive(isPiPActive)
-    pipStateRef.current = isPiPActive
-    
-    // If PIP is activated in wide mode, smoothly scroll to top
-    if (isPiPActive && wideMode) {
-      window.scrollTo({ top: 0, behavior: 'smooth' })
-    }
-  }, [wideMode])
-
-  const handleAutoplayToggle = useCallback(() => {
-    const newMode = !autoplayMode
-    setAutoplayMode(newMode)
-    
-    // Save preference to localStorage
-    try {
-      localStorage.setItem('autoplay_mode', String(newMode))
-    } catch {}
-    
-    toast({
-      title: newMode ? "Auto-play Mode Enabled" : "Auto-play Mode Disabled",
-      description: newMode 
-        ? "Videos will automatically advance to the next chapter" 
-        : "Videos will pause at the end of each chapter",
-    })
-  }, [autoplayMode, toast])
-
-  const handleMobilePlaylistToggle = useCallback(() => {
-    setMobilePlaylistOpen(!mobilePlaylistOpen)
-  }, [mobilePlaylistOpen])
-
-  const handleFullscreenToggle = useCallback(() => {
-    if (!document.fullscreenElement) {
-      document.documentElement.requestFullscreen()
-    } else {
-      document.exitFullscreen()
-    }
-  }, [])
-
-  const handleTheaterModeToggle = useCallback(() => {
-    setIsTheaterMode(!isTheaterMode)
-  }, [isTheaterMode])
-
   const handleVideoLoad = useCallback(() => {
     setIsVideoLoading(false)
   }, [])
@@ -254,117 +156,19 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
 
   const handleVideoProgress = useCallback((state: { playedSeconds: number }) => {
     if (currentVideoId) {
+      const videoStateStore = useVideoState
       videoStateStore.getState().updateProgress(currentVideoId, state.playedSeconds)
     }
-  }, [currentVideoId, videoStateStore])
+  }, [currentVideoId])
 
-  const handleVideoEnd = useCallback(() => {
-    setVideoEnding(true)
-    
-    if (autoplayMode && nextChapter) {
-      setShowChapterTransition(true)
-      setNextChapterInfo({
-        title: nextChapter.chapter.title,
-        description: nextChapter.chapter.description,
-        thumbnail: nextChapter.chapter.thumbnail || undefined,
-        duration: nextChapter.duration,
-      })
-      
-      // Start countdown
-      const countdown = setInterval(() => {
-        setAutoplayCountdown((prev) => {
-          if (prev <= 1) {
-            clearInterval(countdown)
-            handleChapterSelect(nextChapter.chapter)
-            setShowChapterTransition(false)
-            setNextChapterInfo(null)
-            setAutoplayCountdown(5)
-            return 5
-          }
-          return prev - 1
-        })
-      }, 1000)
+  const handleVideoEndWithAutoplay = useCallback(() => {
+    if (nextChapter) {
+      handleVideoEnd(nextChapter, () => handleChapterSelect(nextChapter.chapter))
     } else {
-      setShowAutoplayOverlay(true)
+      // Show certificate when course is complete
+      setTimeout(() => setShowCertificate(true), 1000)
     }
-  }, [autoplayMode, nextChapter, handleChapterSelect])
-
-  const handleChapterComplete = useCallback(async () => {
-    if (!currentChapter?.id || !user) return
-    
-    try {
-      await dispatch(markChapterAsCompleted({
-        chapterId: currentChapter.id.toString(),
-        courseId: course.id.toString(),
-      }))
-      
-      // Update progress
-      if (updateProgress) {
-        updateProgress({
-          courseId: course.id,
-          chapterId: currentChapter.id.toString(),
-          completed: true,
-        })
-      }
-      
-      toast({
-        title: "Chapter Completed!",
-        description: "Great job! You've completed this chapter.",
-      })
-    } catch (error) {
-      console.error("Error marking chapter as completed:", error)
-      toast({
-        title: "Error",
-        description: "Failed to mark chapter as completed. Please try again.",
-        variant: "destructive",
-      })
-    }
-  }, [currentChapter?.id, user, dispatch, course.id, updateProgress, toast])
-
-  const handleChapterSelect = useCallback(async (chapter: FullChapterType) => {
-    if (!chapter?.id) {
-      toast({
-        title: "Invalid Chapter",
-        description: "Please select a valid chapter.",
-        variant: "destructive",
-      })
-      return
-    }
-
-    try {
-      // Validate chapter has video
-      if (!chapter.videoId) {
-        toast({
-          title: "Video Unavailable",
-          description: "This chapter doesn't have a video available.",
-          variant: "destructive",
-        })
-        return
-      }
-
-      // Update Redux state
-      dispatch(setCurrentVideoApi(chapter.videoId))
-
-      // Update Zustand store
-      videoStateStore.getState().setCurrentVideo(chapter.videoId, course.id)
-
-      // Close mobile playlist if open
-      setMobilePlaylistOpen(false)
-
-      // Show success feedback
-      toast({
-        title: "Chapter Selected",
-        description: `Now playing: ${chapter.title}`,
-      })
-    } catch (error) {
-      console.error("Error selecting chapter:", error)
-      toast({
-        title: "Error",
-        description: "Failed to select chapter. Please try again.",
-        variant: "destructive",
-      })
-    }
-  }, [course.id, dispatch, toast, videoStateStore, setMobilePlaylistOpen])
+  }, [nextChapter, handleVideoEnd, handleChapterSelect])
 
   const handleNextVideo = useCallback(() => {
     if (!currentVideoId || !nextChapter) return
@@ -376,115 +180,18 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
     handleChapterSelect(prevChapter.chapter)
   }, [currentVideoId, prevChapter, handleChapterSelect])
 
-  const handleCancelAutoplay = useCallback(() => {
-    setShowChapterTransition(false)
-    setShowAutoplayOverlay(false)
-    setNextChapterInfo(null)
-    setAutoplayCountdown(5)
-  }, [])
+  const handleMobilePlaylistToggle = useCallback(() => {
+    setMobilePlaylistOpen(!mobilePlaylistOpen)
+  }, [mobilePlaylistOpen])
 
   const handleCertificateClick = useCallback(() => {
     setShowCertificate(true)
   }, [])
 
-  const handleWideModeToggle = useCallback(() => {
-    const newWideMode = !wideMode
-    setWideMode(newWideMode)
-    
-    // Save preference to localStorage
-    try {
-      localStorage.setItem(`wide_mode_course_${course.id}`, String(newWideMode))
-    } catch {}
-    
-    // Show feedback toast
-    toast({
-      title: newWideMode ? "Wide Mode Enabled" : "Wide Mode Disabled",
-      description: newWideMode 
-        ? "Video player now uses full width" 
-        : "Video player now uses standard width",
-    })
-  }, [wideMode, course.id, toast])
-
   // ============================================================================
-  // EFFECTS - ALL DEPENDENCIES MUST BE PROPERLY DECLARED
+  // EFFECTS
   // ============================================================================
   
-  // Load wide mode preference from localStorage
-  useEffect(() => {
-    try {
-      const savedWideMode = localStorage.getItem(`wide_mode_course_${course.id}`)
-      if (savedWideMode !== null) {
-        setWideMode(savedWideMode === 'true')
-      }
-    } catch {}
-  }, [course.id])
-
-  // Load autoplay mode preference from localStorage
-  useEffect(() => {
-    try {
-      const savedAutoplayMode = localStorage.getItem('autoplay_mode')
-      if (savedAutoplayMode !== null) {
-        setAutoplayMode(savedAutoplayMode === 'true')
-      }
-    } catch {}
-  }, [])
-
-  // Listen for fullscreen changes
-  useEffect(() => {
-    const handleFullscreenChange = () => {
-      setIsFullscreen(!!document.fullscreenElement)
-    }
-
-    document.addEventListener('fullscreenchange', handleFullscreenChange)
-    return () => document.removeEventListener('fullscreenchange', handleFullscreenChange)
-  }, [])
-
-  // ESC key handler for exiting modes
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      // Only handle shortcuts when not typing in input fields
-      if (event.target instanceof HTMLInputElement || event.target instanceof HTMLTextAreaElement) {
-        return
-      }
-
-      switch (event.key) {
-        case 'Escape':
-          if (showChapterTransition) {
-            handleCancelAutoplay()
-          } else if (showAutoplayOverlay) {
-            setShowAutoplayOverlay(false)
-          } else if (isPiPActive) {
-            handlePIPToggle(false)
-          }
-          break
-        case 'p':
-        case 'P':
-          if (event.ctrlKey || event.metaKey) {
-            event.preventDefault()
-            handlePIPToggle(!isPiPActive)
-          }
-          break
-        case 'a':
-        case 'A':
-          if (event.ctrlKey || event.metaKey) {
-            event.preventDefault()
-            handleAutoplayToggle()
-          }
-          break
-      }
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [
-    showChapterTransition, 
-    showAutoplayOverlay, 
-    isPiPActive, 
-    handleCancelAutoplay, 
-    handlePIPToggle, 
-    handleAutoplayToggle
-  ])
-
   // Enhanced initialization logic for video selection
   useEffect(() => {
     if (videoPlaylist.length === 0) {
@@ -519,30 +226,27 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
       dispatch(setCurrentVideoApi(targetVideo.videoId))
 
       // Update Zustand store
+      const videoStateStore = useVideoState
       videoStateStore.getState().setCurrentVideo(targetVideo.videoId, course.id)
 
       console.log(`[MainContent] Initialized with video: ${targetVideo.videoId}, course: ${course.id}`)
     } else {
       console.error("Failed to select a video")
     }
-  }, [course.id, initialChapterId, videoPlaylist, dispatch, videoStateStore, currentVideoId, progress])
+  }, [course.id, initialChapterId, videoPlaylist, dispatch, currentVideoId, progress])
 
-  // Resume prompt
-  useEffect(() => {
-    if (progress && !resumePromptShown && progress.currentChapterId && !currentVideoId && user) {
-      const resumeChapter = videoPlaylist.find(
-        (entry) => entry.chapter.id.toString() === progress.currentChapterId?.toString(),
-      )
-
-      if (resumeChapter) {
-        setResumePromptShown(true)
-        toast({
-          title: "Resume Learning",
-          description: `Continue from "${resumeChapter.chapter.title}"? (Resume feature available in your dashboard)`
-        })
-      }
-    }
-  }, [progress, resumePromptShown, currentVideoId, videoPlaylist, toast, user])
+  // ============================================================================
+  // KEYBOARD SHORTCUTS
+  // ============================================================================
+  
+  useKeyboardShortcuts({
+    showChapterTransition,
+    showAutoplayOverlay,
+    isPiPActive,
+    onCancelAutoplay: handleCancelAutoplay,
+    onPIPToggle: handlePIPToggle,
+    onAutoplayToggle: handleAutoplayToggle
+  })
 
   // ============================================================================
   // RENDER LOGIC
@@ -578,7 +282,7 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
     onPlayerReady: handlePlayerReady,
     onSeekToBookmark: handleSeekToBookmark,
     onVideoProgress: handleVideoProgress,
-    onVideoEnd: handleVideoEnd,
+    onVideoEnd: handleVideoEndWithAutoplay,
     onPictureInPictureToggle: handlePIPToggle,
     onFullscreenToggle: handleFullscreenToggle,
     isPiPActive,
@@ -590,7 +294,7 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
     handlePlayerReady, 
     handleSeekToBookmark, 
     handleVideoProgress, 
-    handleVideoEnd, 
+    handleVideoEndWithAutoplay, 
     handlePIPToggle, 
     handleFullscreenToggle, 
     isPiPActive
@@ -599,78 +303,16 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
   return (
     <div className="min-h-screen bg-background">
       {/* Top Toolbar */}
-      <div className="sticky top-0 z-40 bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 border-b">
-        <div className="flex items-center justify-between p-4">
-          <div className="flex items-center space-x-4">
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => router.back()}
-              className="flex items-center space-x-2"
-            >
-              <ChevronLeft className="h-4 w-4" />
-              <span>Back</span>
-            </Button>
-            
-            <div className="flex items-center space-x-2">
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={handleWideModeToggle}
-                className={cn(
-                  "flex items-center space-x-2",
-                  wideMode && "bg-primary/10 text-primary"
-                )}
-              >
-                <Maximize2 className="h-4 w-4" />
-                <span>Wide Mode</span>
-              </Button>
-              
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={handlePIPToggle.bind(null, !isPiPActive)}
-                className={cn(
-                  "flex items-center space-x-2",
-                  isPiPActive && "bg-primary/10 text-primary"
-                )}
-              >
-                <PictureInPicture className="h-4 w-4" />
-                <span>PIP</span>
-              </Button>
-              
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={handleAutoplayToggle}
-                className={cn(
-                  "flex items-center space-x-2",
-                  autoplayMode && "bg-primary/10 text-primary"
-                )}
-              >
-                <Play className="h-4 w-4" />
-                <span>Auto-play</span>
-              </Button>
-            </div>
-          </div>
-          
-          <div className="flex items-center space-x-2">
-            {isPiPActive && (
-              <div className="px-3 py-1 bg-primary/10 text-primary text-sm rounded-full">
-                PIP Active
-              </div>
-            )}
-            
-            {autoplayMode && (
-              <div className="px-3 py-1 bg-green-500/10 text-green-600 text-sm rounded-full">
-                Auto-play Mode
-              </div>
-            )}
-            
-            <CourseActions course={course} isOwner={isOwner} />
-          </div>
-        </div>
-      </div>
+      <CourseToolbar
+        course={course}
+        isOwner={isOwner}
+        wideMode={wideMode}
+        isPiPActive={isPiPActive}
+        autoplayMode={autoplayMode}
+        onWideModeToggle={handleWideModeToggle}
+        onPIPToggle={handlePIPToggle}
+        onAutoplayToggle={handleAutoplayToggle}
+      />
 
       {/* Main Content Grid */}
       <div className={cn("grid gap-6 p-6", gridLayoutClasses)}>
@@ -733,19 +375,18 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
           >
             {isPiPActive ? (
               <div className="text-center p-8 text-muted-foreground">
-                <PictureInPicture className="h-12 w-12 mx-auto mb-4 opacity-50" />
+                <div className="text-6xl mb-4">📺</div>
                 <h3 className="text-lg font-semibold mb-2">Picture-in-Picture Mode</h3>
                 <p className="text-sm mb-4">
                   Video is now playing in a floating window. 
                   Close PIP to return to normal view.
                 </p>
-                <Button
-                  variant="outline"
-                  size="sm"
+                <button
                   onClick={() => handlePIPToggle(false)}
+                  className="px-4 py-2 bg-primary text-primary-foreground rounded-md hover:bg-primary/90 transition-colors"
                 >
                   Exit PIP Mode
-                </Button>
+                </button>
               </div>
             ) : (
               <MemoizedVideoNavigationSidebar {...sidebarProps} />
@@ -756,13 +397,14 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
 
       {/* Mobile Playlist Toggle */}
       <div className="md:hidden fixed bottom-4 right-4 z-50">
-        <Button
+        <button
           onClick={handleMobilePlaylistToggle}
-          size="lg"
-          className="rounded-full w-14 h-14 shadow-lg"
+          className="w-14 h-14 bg-primary text-primary-foreground rounded-full shadow-lg flex items-center justify-center hover:bg-primary/90 transition-colors"
         >
-          <Play className="h-6 w-6" />
-        </Button>
+          <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14.828 14.828a4 4 0 01-5.656 0M9 10h1m4 0h1m-6 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+          </svg>
+        </button>
       </div>
 
       {/* Mobile Playlist Overlay */}
@@ -777,13 +419,14 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
             <div className="flex flex-col h-full">
               <div className="flex items-center justify-between p-4 border-b">
                 <h3 className="text-lg font-semibold">Course Chapters</h3>
-                <Button
-                  variant="ghost"
-                  size="sm"
+                <button
                   onClick={handleMobilePlaylistToggle}
+                  className="p-2 hover:bg-muted rounded-md"
                 >
-                  <ChevronLeft className="h-4 w-4" />
-                </Button>
+                  <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
+                  </svg>
+                </button>
               </div>
               
               <div className="flex-1 overflow-y-auto p-4">
@@ -826,13 +469,12 @@ const MainContent: React.FC<ModernCoursePageProps> = ({
                 )}
               </div>
               
-              <Button
-                variant="outline"
+              <button
                 onClick={handleCancelAutoplay}
-                className="w-full"
+                className="w-full px-4 py-2 bg-muted text-muted-foreground rounded-md hover:bg-muted/80 transition-colors"
               >
                 Cancel Auto-play
-              </Button>
+              </button>
             </div>
           </motion.div>
         )}
